requirements.txt
annotated-types==0.7.0
anyio==4.10.0
asyncpg==0.30.0
bcrypt==4.3.0
black==25.9.0
certifi==2025.8.3
cffi==2.0.0
chardet==5.2.0
charset-normalizer==3.4.3
click==8.3.0
coverage==7.10.6
cryptography==46.0.1
deprecation==2.1.0
distro==1.9.0
ecdsa==0.19.1
fastapi==0.117.1
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
iniconfig==2.1.0
jiter==0.11.0
lxml==6.0.1
mypy==1.18.2
mypy_extensions==1.1.0
numpy==2.3.3
openai==1.108.1
packaging==25.0
passlib==1.7.4
pathspec==0.12.1
pdfminer.six==20250506
pdfplumber==0.11.7
pgvector==0.4.1
pillow==11.3.0
platformdirs==4.4.0
pluggy==1.6.0
postgrest==2.19.0
prometheus_client==0.23.1
pyasn1==0.6.1
pycparser==2.23
pydantic==2.11.9
pydantic-settings==2.10.1
pydantic_core==2.33.2
Pygments==2.19.2
PyJWT==2.10.1
pypandoc==1.15
PyPDF2==3.0.1
pypdfium2==4.30.0
pytest==8.4.2
pytest-asyncio==1.2.0
pytest-cov==7.0.0
pytest-mock==3.15.1
python-docx==1.2.0
python-dotenv==1.1.1
python-jose==3.5.0
python-multipart==0.0.20
pytokens==0.1.10
PyYAML==6.0.2
realtime==2.19.0
regex==2025.9.18
requests==2.32.5
rsa==4.9.1
ruff==0.13.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.43
starlette==0.48.0
storage3==2.19.0
StrEnum==0.4.15
structlog==25.4.0
supabase==2.19.0
supabase-auth==2.19.0
supabase-functions==2.19.0
tenacity==9.1.2
tiktoken==0.11.0
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.36.0
uvloop==0.21.0
watchfiles==1.1.0
websockets==15.0.1
psutil==6.1.1
